[quarkus] add tests for jvm only loaders #215
