+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-core/src/main/resources/application.properties
@@ -14,14 +14,8 @@
 ## See the License for the specific language governing permissions and
 ## limitations under the License.
 ## ---------------------------------------------------------------------------
+
 #
 # Quarkus
 #
-quarkus.log.file.enable    = false
-quarkus.log.console.enable = false
 quarkus.banner.enabled     = false
-
-#
-# Camel
-#
-camel.context.name = camel-k-quarkus-ctx
