+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-cron/src/main/resources/application.properties
@@ -14,11 +14,10 @@
 ## See the License for the specific language governing permissions and
 ## limitations under the License.
 ## ---------------------------------------------------------------------------
+
 #
 # Quarkus
 #
-quarkus.log.file.enable    = false
-quarkus.log.console.enable = true
 quarkus.banner.enabled     = false
 
 #
@@ -26,11 +25,3 @@ quarkus.banner.enabled     = false
 #
 quarkus.camel.main.routes-discovery.enabled = false
 
-#
-# Camel
-#
-camel.context.name = camel-k-quarkus-ctx
-
-#
-# Integration
-#
