+++ camel-k-quarkus/camel-k-quarkus-itests/pom.xml
@@ -32,6 +32,9 @@
         <module>camel-k-quarkus-itests-core</module>
         <module>camel-k-quarkus-itests-cron</module>
         <module>camel-k-quarkus-itests-knative</module>
+        <module>camel-k-quarkus-itests-loader-support</module>
+        <module>camel-k-quarkus-itests-loader-groovy</module>
+        <module>camel-k-quarkus-itests-loader-java</module>
         <module>camel-k-quarkus-itests-loader-js</module>
         <module>camel-k-quarkus-itests-loader-xml</module>
         <module>camel-k-quarkus-itests-loader-yaml</module>
